Simplifications in the GraphQLGenerator class
This is the first change of a series of changes to redesign the GraphQL generation service as a more functional module. The current implementation is becoming hard to understand because we're starting to mix the implementations of different kinds of GraphQL queries and overusig injections. We should aim for a more modular design in which functionality is properly isolated and untangled and root classes don't have knowledge of implementation details of leaf classes. For instance, if we differentiate between regular or paginated queries, it would be desirable that we have separate generators for each. This will make it easier to change, improve and test these classes. This first change cleans a small amount of redundancy from the `GraphQLGenerator` class by avoiding having an explicit singleton instance. In upcoming changes, I would like to make the `GraphQLGenerator` root class less aware of what the other classes are doing.